requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
elasticsearch==8.8.0
python-dotenv==1.0.0
pydantic==1.10.8